chore : 병원리스트2였던 뷰파일로 라우팅 설정, 스타일 수정
